requirements.txt
tweepy==4.14.0
textblob==0.17.1
plotly==5.18.0
pandas==2.1.4
streamlit==1.29.0
python-dotenv==1.0.0